Add end to end save code now check
This defines an end to end check which trigger for an dedicated origin with a given type
a save code now requests.

This then regularly polls the status out of the save code now api. Depending on the time
it took to actually ingest and the ingestion status, it will return the status of said
ingestion and roughly the time it took.

The end goal being to deploy such checks into icinga and being notified when a save code
now request takes too long.

Example use:
```
$ swh icinga_plugins \
  check-savecodenow --swh-storage-url http://localhost:5002/ \
  --swh-web-url https://webapp.staging.swh.network \
  origin <url> \
  --visit-type git

SAVECODENOW WARNING - Save code now request for ('git', '<url>') took 30.58s and succeeded.
| 'total_time' = 30.58s
```

Related to T2117