deposit: Deal with rate limit during get queries
We currently have some icing alerts about 429 response [1]. The current code is not able
to deal with it. So, this reuses the retry decorator used in some other parts of the swh
base code to retry with a backoff when this occurs. This should fix the current
failures.

[1]
```
DEPOSIT CRITICAL - Getting the list of metadata returned code 429:
b'{"exception":"Throttled",​"reason":"Request was throttled. Expected available in 16
seconds."}'
```

Refs. swh/infra/sysadm-environment#5129